Extract tree rendering
This change extracts the tree rendering part of Radiography into a separate class, paving the way for supporting other trees.

TreeStringRenderer does the heavy lifting of navigating the tree and printing the hierarchy. Note that it's not a nice functional / lazy API. This change sticks to the approach in Radiography to limit temporary object creation.